+++ packages/eslint-plugin-code-policy/CHANGELOG.md
@@ -1,6 +1,6 @@
 # eslint-plugin-code-policy
 
-## Unreleased
+## 0.7.0
 
 ### Minor Changes